# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API bitkakir project",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/banned-terms": {
            "get": {
                "tags": [
                    "Banned Terms"
                ],
                "summary": "List banned-terms",
                "description": "Display listing of the resource",
                "operationId": "cf13f8f0f1e1c57feafd329feb0033b9",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Banned Terms"
                ],
                "summary": "Save banned terms",
                "description": "Creates and returns a new banned terms",
                "operationId": "61a2ffc0e14729b48154a03118409ac4",
                "requestBody": {
                    "$ref": "#/components/requestBodies/BannedTermRequest"
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
...
        "/api/v1/company-data": {
            "get": {
                "tags": [
                    "Company data"
                ],
                "summary": "List company data",
                "description": "Display listing of the resource",
                "operationId": "092e5b116d10a0a29b72ca126ff56e83",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Company data"
                ],
                "summary": "Save company data",
                "description": "Creates and returns a new company data",
                "operationId": "a48076b48315b73daf7d4ded19751698",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CompanyDataRequest"
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
...
        "/api/v1/banned-terms/{id}": {
            "get": {
                "tags": [
                    "Banned Terms"
                ],
                "summary": "Show banned terms by ID",
                "description": "Display a listing by ID",
                "operationId": "d37bb96a17c786524013a99ecb3cd7c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Banned Terms"
                ],
                "summary": "Update banned terms",
                "description": "Update the specified resource in storage",
                "operationId": "7af1878969a369f6cf655ed849324afc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/BannedTermRequest"
                },
                "responses": {
                    "201": {
                        "description": "Category updated successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Banned Terms"
                ],
                "summary": "Remove the specified resource from storage",
                "description": "Delete banned terms",
                "operationId": "13511b8cd925398ec33b1e9df4bd106e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/v1/company-data/{id}": {
            "get": {
                "tags": [
                    "Company data"
                ],
                "summary": "Show company data by UUID or ID",
                "description": "Display a listing by UUID or ID",
                "operationId": "290889394d7eb559603f9f9bf6fa3ffd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by UUID or ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Company data"
                ],
                "summary": "Update company data",
                "description": "Update the specified resource in storage",
                "operationId": "372a82324b505b96f8051ec1888f8125",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by UUID or ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CompanyDataRequest"
                },
                "responses": {
                    "201": {
                        "description": "Product updated successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company data"
                ],
                "summary": "Remove the specified resource from storage",
                "description": "Delete company data",
                "operationId": "00b66be1a31c3eef125403b65606e75f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete by UUID or ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/v1/company-data/updateMobile": {
            "post": {
                "tags": [
                    "Company data"
                ],
                "summary": "Update company data",
                "description": "Update the specified resource in storage",
                "operationId": "03dc9e6637bd96ed820e51f1fbe8c324",
                "requestBody": {
                    "description": "Informações para registrar uma empresa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "companyName",
                                    "contact"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "companyName": {
                                        "type": "string",
                                        "example": "XPTO Angola Lda"
                                    },
                                    "nif": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "personalNif": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua do Comércio, Luanda"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Tecnologia"
                                    },
                                    "contactMethod": {
                                        "type": "string",
                                        "example": "whatsapp"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": "+244912345678"
                                    },
                                    "hasSite": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "example": "https://empresa.co.ao"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product updated successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
...
    },
    "components": {
        "requestBodies": {
            "BannedTermRequest": {
                "description": "Texto proibido que será cadastrado",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "text_en"
                            ],
                            "properties": {
                                "text_en": {
                                    "type": "string",
                                    "example": "badword"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "CompanyDataRequest": {
                "description": "Dados da empresa",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "nif",
                                "user_id"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Empresa XPTO Ltda"
                                },
                                "nif": {
                                    "type": "string",
                                    "example": "123456789"
                                },
                                "location": {
                                    "type": "string",
                                    "example": "Luanda, Angola"
                                },
                                "certification": {
                                    "type": "string",
                                    "example": "ISO 9001"
                                },
                                "user_id": {
                                    "type": "string",
                                    "example": "abc123xyz"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}